Add packages to fields in init_atmosphere core Registry based on output packages
In general, for fields that are written to the default "output" stream by the
init_atmosphere core, those fields only need to be allocated if they are active
in the "output" stream. This implies that we can apply packages from the
"output" stream to the fields themselves to potentially save memory (depending
on which pre-processing stages are active). Note that there exist fields -- for
example, first-guess fields -- that are not written to the "output" stream but
that still must be allocated during some pre-processing stages.